% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacheManager.R
\name{cacheManager}
\alias{cacheManager}
\title{Manage the cache}
\usage{
cacheManager(nameFile)
}
\arguments{
\item{nameFile}{the name of the file or folder}
}
\value{
the path to the file we need
}
\description{
The function creates a folder (only once) and then controls files that need to be downloaded and
placed in this folder.
}
\section{Localisation}{

The localisation of the folder is :
  - On Linux : `~/.local/share/R/osmose`
  - On Mac OS X : `~/Library/Application Support/R/osmose`
  - On Windows 7 up to 10 : `C:\\Users\\<username>\\AppData\\Local\\R\\BIOMASS\\R\\osmose`
  - On Windows XP : `C:\\Documents and Settings\\<username>\\Data\\R\\BIOMASS\\R\\osmose`

See this function for more information : [rappdirs::user_data_dir()][BIOMASS::cacheManager(nameFile)]
}

\seealso{
[rappdirs::user_data_dir()][BIOMASS::cacheManager(nameFile)]
}
\author{
Arthur PERE

Nicolas BARRIER
}
\keyword{Internal}
