% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose-internal.R
\name{configureCalibration}
\alias{configureCalibration}
\title{Reads calibration parameters from an osmose.config list.}
\usage{
configureCalibration(L1)
}
\arguments{
\item{L1}{osmose.config object (see \code{\link{readOsmoseConfiguration}})}
}
\value{
A list of parameters to calibrate ("guess", "max", "min", "phase")
}
\description{
The configuration argument must contain a "calibration" entry to work.
}
