% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose.netcdf.R
\name{correct_ltl_file}
\alias{correct_ltl_file}
\title{Creates a NetCDF file containing the new format for LTL
concentrations. The "ltl_biomass" of dims (ntime, nltl, nlat, nlon)
is splitted into nltl variables of dims (ntime, nlat, nlon)}
\usage{
correct_ltl_file(ltl_filename, osmose_config, varlon = "longitude",
  varlat = "latitude", vartime = "time", absolute = TRUE)
}
\arguments{
\item{ltl_filename}{Name of the NetCDF LTL file}

\item{osmose_config}{Name of the main configuration file}

\item{varlon}{Name of the longitude variable}

\item{varlat}{Name of the latitude variable}

\item{vartime}{Name of the time variable}

\item{absolute}{Whether the path is absolute (TRUE) or relative (FALSE)}
}
\description{
Creates a NetCDF file containing the new format for LTL
concentrations. The "ltl_biomass" of dims (ntime, nltl, nlat, nlon)
is splitted into nltl variables of dims (ntime, nlat, nlon)
}
