% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose-plot-util.R
\name{.osmose.barplot}
\alias{.osmose.barplot}
\title{Plots a barplot, with xlabels rotated with a 45degree angle}
\usage{
.osmose.barplot(x, add_text = TRUE, color = NULL, ...)
}
\arguments{
\item{x}{Data array}

\item{add_text}{True if text should be added.}

\item{color}{Color of the barplot}

\item{...}{Additional arguments to the barplot function (color, etc.)}
}
\description{
Plots a barplot, with xlabels rotated with a 45degree angle
}
