% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose-plot-util.R
\name{.osmose.format_data_stacked}
\alias{.osmose.format_data_stacked}
\title{Reformat the data into the ggplot2 stacked plot format.
It returns a dataframe with specie, time and predation rate
in the columns.}
\usage{
.osmose.format_data_stacked(data, time = NULL)
}
\arguments{
\item{data}{Data to plot}

\item{time}{Time vector}
}
\description{
Reformat the data into the ggplot2 stacked plot format.
It returns a dataframe with specie, time and predation rate
in the columns.
}
