% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose-plot-util.R
\name{.plot_osmose_dietMatrix}
\alias{.plot_osmose_dietMatrix}
\title{Plots diet matrix}
\usage{
.plot_osmose_dietMatrix(
  x,
  time.mean,
  species,
  thres,
  color,
  add_text,
  plot_name,
  legsize = 1,
  ...
)
}
\arguments{
\item{x}{Diet matrix}

\item{time.mean}{If TRUE, the time mean diet matrix is computed is displayed.}

\item{species}{Species name}

\item{thres}{Thresholds (in percentage). Time-average predation rates below this threashold 
are binned together ("other" column).}

\item{color}{Color of the barplot.}

\item{add_text}{\code{TRUE} if text should be added.}

\item{plot_name}{Label for ylab in diet matrix plot.}

\item{legsize}{Size of the legend (default 1), if time.mean is \code{FALSE}.}

\item{...}{Additional plot arguments passed to barplot.}
}
\value{
None
}
\description{
Plots diet matrix
}
