% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose-internal.R
\name{.read_osmose_ncdf}
\alias{.read_osmose_ncdf}
\title{Function to read osmose netcdf files}
\usage{
.read_osmose_ncdf(files, path, varid, calendar = 360, ...)
}
\arguments{
\item{files}{String of name of the file that will be read.}

\item{path}{String of path of the file that will be read}

\item{...}{Extra arguments}
}
\description{
Function to read osmose netcdf files
}
