% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose-internal.R
\name{.readOsmoseCsv}
\alias{.readOsmoseCsv}
\title{Read Osmose CSV files}
\usage{
.readOsmoseCsv(
  file,
  sep = ",",
  skip = 1,
  row.names = 1,
  na.strings = c("NA", "NaN"),
  rm = 1,
  ...
)
}
\arguments{
\item{file}{File name}

\item{sep}{Column separator}

\item{skip}{Number of line to skip}

\item{row.names}{Index of the row names column}

\item{na.strings}{List of NaN strings}

\item{rm}{???? Not used}

\item{...}{Extra arguments passed to the method.}
}
\value{
A data frame
}
\description{
Read Osmose CSV files
}
