% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose-internal.R
\name{existOsmoseParameter}
\alias{existOsmoseParameter}
\title{Check if a parameter exists}
\usage{
existOsmoseParameter(par, ..., keep.att = FALSE)
}
\arguments{
\item{par}{Output of the \code{link{readOsmoseConfiguration}} function}

\item{...}{String arguments}

\item{keep.att}{Whether parameter attributes should be kept}
}
\description{
Check if a parameter exists
}
\examples{
{
   filename = system.file("extdata", "gog/osm_all-parameters.csv", package="osmose")
   par = readOsmoseConfiguration(filename)
   getOsmoseParameter(par, "population", "seeding", "year", "max", keep.att=FALSE)
}
}
