% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose.netcdf.R
\name{extract_grid_mask}
\alias{extract_grid_mask}
\title{Extracts the model grid file from the LTL forcing file.}
\usage{
extract_grid_mask(
  filename,
  varname = "ltl_biomass",
  output_file = "ltl-grid-mask.csv",
  fillVal = NA,
  ...
)
}
\arguments{
\item{filename}{LTL forcing file name}

\item{varname}{Name of the LTL biomass variable (default "ltl_biomass")}

\item{output_file}{Name of the output file (default "ltl-grid-mask.csv")}

\item{fillVal}{Fill Value (default NA but could be 0).}

\item{...}{Additional arguments of the \code{ncvar_get} function. See Details.}
}
\description{
Extracts the model grid file from the LTL forcing file.
}
\details{
\code{...} can be used to allow more flexibility on the choice of 
the mask variable via the start and count arguments in order to get a final
2D matrix.
}
