% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose.netcdf.R
\name{extract_grid_param}
\alias{extract_grid_param}
\title{Computes the Osmose grid parameters by using the
LTL forcing file. The parameters are prompted in the console.
The user should copy the console output and paste it into
the Osmose configuration file.}
\usage{
extract_grid_param(
  filename,
  lonname = "longitude",
  latname = "latitude",
  outfile = NULL
)
}
\arguments{
\item{filename}{NetCDF filename}

\item{lonname}{Longitude name (default "longitude")}

\item{latname}{Latitude name (default "latitude")}

\item{outfile}{Output CSV file (if NULL, no file is written). The directory is created recursively
if it does not exist
Warning: This function must be checked in order to be sure that it's not just
being created for an specific version of osmose-java (the return has names of
parameters that are only valid for an specific version). Besides, it'd be 
better to create a more global function that extract parameters, mask, etc 
from LTL netCDF files ('extract_grid_mask' function).}
}
\description{
Computes the Osmose grid parameters by using the
LTL forcing file. The parameters are prompted in the console.
The user should copy the console output and paste it into
the Osmose configuration file.
}
