% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose.netcdf.R
\name{find_species_index}
\alias{find_species_index}
\title{Returns the index of a species within the configuration.}
\usage{
find_species_index(spec, param)
}
\arguments{
\item{spec}{Name of the species}

\item{param}{Osmose configuration (osmose.config object)}
}
\value{
An integer containing the species index
}
\description{
Returns the index of a species within the configuration.
}
