% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacheManager.R
\name{flushCache}
\alias{flushCache}
\title{#' @keywords Internal
checkTime <- function() {
  # Check if it's time to update
  if (!file.exists(cachePath("last_check.txt"))) {
    writeLines(as.character(Sys.Date()), cachePath("last_check.txt"))
  } else {
    check_interval <- 1830 # this message will appear every 5 years
    
    last_check <- as.Date(readLines(cachePath("last_check.txt")))
    if ((Sys.Date() - last_check) > check_interval) {
      message(
        "You can verify if the cache is updated by using this function\n",
        "\t\tupdateCache()\n",
        "Be careful, all the environement variable will be deleted and updated.\n",
        "You can ignore this message, and can prevent this message to appear again by using\n",
        "\t\toptions(BIOMASS.ignore_update=TRUE)"
      )
      
      # update the flag
      writeLines(as.character(Sys.Date()), cachePath("last_check.txt"))
    }
  }
  
  return()
}}
\usage{
flushCache()
}
\description{
#' @keywords Internal
checkTime <- function() {
  # Check if it's time to update
  if (!file.exists(cachePath("last_check.txt"))) {
    writeLines(as.character(Sys.Date()), cachePath("last_check.txt"))
  } else {
    check_interval <- 1830 # this message will appear every 5 years
    
    last_check <- as.Date(readLines(cachePath("last_check.txt")))
    if ((Sys.Date() - last_check) > check_interval) {
      message(
        "You can verify if the cache is updated by using this function\n",
        "\t\tupdateCache()\n",
        "Be careful, all the environement variable will be deleted and updated.\n",
        "You can ignore this message, and can prevent this message to appear again by using\n",
        "\t\toptions(BIOMASS.ignore_update=TRUE)"
      )
      
      # update the flag
      writeLines(as.character(Sys.Date()), cachePath("last_check.txt"))
    }
  }
  
  return()
}
}
\keyword{Internal}
