% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose-parameter_modeling.R
\name{gaussian_dispersal}
\alias{gaussian_dispersal}
\title{Generate a migration input file using gaussian dispersal.}
\usage{
gaussian_dispersal(par, conf, sp)
}
\arguments{
\item{par}{A list with parameters, see detail.}

\item{conf}{The OSMOSE configuration file}

\item{sp}{The species code.}
}
\value{
A data.frame to be written with write_osmose()
}
\description{
Generate a migration input file using gaussian dispersal.
}
