% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose.config-class.R
\name{get_var.osmose.configuration}
\alias{get_var.osmose.configuration}
\title{get_var method for osmose configuration objects}
\usage{
\method{get_var}{osmose.configuration}(object, what, ...)
}
\arguments{
\item{object}{Object of \code{osmose.config} class. 
See the \code{\link{read_osmose}} and \code{\link{readOsmoseConfiguration}} functions
 for more information about this object.}

\item{what}{Name of the variable to extract from the configuration file.}

\item{...}{Extra arguments for plotting method.}
}
\value{
An object of \code{list} class containing all the relevant information about 
the variable extracted.
}
\description{
Get the configuration files from
}
