% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose.interannual-fishing.R
\name{getFishingBaseRate.byDt}
\alias{getFishingBaseRate.byDt}
\title{Get fishing base rate using time-step rates.}
\usage{
getFishingBaseRate.byDt(sp, fishing, T, ndt)
}
\arguments{
\item{sp}{Current specie (sp0, sp1, etc.)}

\item{fishing}{Fishing parameters}

\item{T}{Number of years}

\item{ndt}{Time step}
}
\description{
It reads the time-step values either from an external file or from the configuration file.
It must contain a number of elements which is a multiple of the fishing period.

Warning: The number of rate elements must be a multiple of the fishing 
period since \strong{recycling} is performed!)
}
