% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose.interannual-fishing.R
\name{getFishingBaseRate.constant}
\alias{getFishingBaseRate.constant}
\title{Get fishing base rate using constant rate}
\usage{
getFishingBaseRate.constant(sp, fishing, T, ndt)
}
\arguments{
\item{sp}{Current specie (sp0, sp1, etc.)}

\item{fishing}{Fishing parameters}

\item{T}{Number of years}

\item{ndt}{Time step}
}
\description{
Get fishing base rate using constant rate
}
\details{
It assumes that in the Osmose configuration, there is a "fishing" entry. It reads
the \emph{rate.spX} parameter, which is repeated \eqn{T \times ndt} times.
}
