% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose.interannual-fishing.R
\name{getFishingBaseRate.linear}
\alias{getFishingBaseRate.linear}
\title{Get fishing base rate using a linear rate}
\usage{
getFishingBaseRate.linear(sp, fishing, T, ndt)
}
\arguments{
\item{sp}{Current specie (sp0, sp1, etc.)}

\item{fishing}{Fishing parameters}

\item{T}{Number of years}

\item{ndt}{Time step}
}
\description{
Get fishing base rate using a linear rate
}
\details{
It assumes that in the Osmose configuration, there is a "fishing" entry. It reads
the \emph{rate.spX} and \emph{rate.slope.spX} parameters. 
The slope is computed any time the fishing frequency changes (fishingperiod parameter)
For instance, if ndt=20, period=5, there are 4 time steps between two fishing time step. Therefore, 
the value will be (i0, i0, i0, i0, i1, i1, i1, i1, ...)
}
