% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose.config-plot-auxiliar.R
\name{getGrowthParameters}
\alias{getGrowthParameters}
\title{Title}
\usage{
getGrowthParameters(par, sp)
}
\arguments{
\item{par}{TODO}

\item{sp}{TODO}
}
\value{
TODO
}
\description{
Title
}
