% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose.config-class.R
\name{getVar.osmose.config}
\alias{getVar.osmose.config}
\title{GetVar method for osmose configuration objects}
\usage{
\method{getVar}{osmose.config}(object, what)
}
\arguments{
\item{object}{Object of \code{osmose.config} class. 
See the \code{\link{read_osmose}} and \code{\link{readOsmoseConfiguration}} functions
 for more information about this object.}

\item{what}{Name of the variable to extract from the configuration file.}
}
\value{
An object of \code{list} class containing all the relevant information about 
the variable extracted.
}
\description{
Get the configuration files from
}
