% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose-class.R
\name{getVar.osmose}
\alias{getVar.osmose}
\title{Get a specific Osmose output.}
\usage{
\method{getVar}{osmose}(object, var, type = "global", expected = FALSE)
}
\arguments{
\item{object}{Osmose outputs (see the \code{\link{osmose2R}} function)}

\item{var}{Variale to extract}

\item{type}{Type of the variable to extract}

\item{expected}{TRUE if the average over the 
last dimensions should be performed if the output is an array.}
}
\value{
An array or a list containing the data.
}
\description{
Get a specific Osmose output.
}
