% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose-class.R
\name{getVar}
\alias{getVar}
\title{Get a specific Osmose output.}
\usage{
getVar(object, var)
}
\arguments{
\item{object}{Osmose outputs (see the \code{\link{osmose2R}} function)}

\item{var}{Variable to extract}
}
\value{
An array or a list containing the data.
}
\description{
Get a specific Osmose output.
}
