% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose-auxiliar.R
\name{isFALSE}
\alias{isFALSE}
\title{Defines isFALSE from isTRUE}
\usage{
isFALSE(x)
}
\arguments{
\item{x}{Condition}
}
\value{
True if condition is not met
}
\description{
Defines isFALSE from isTRUE
}
