% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose-auxiliar.R
\name{list_osmose_versions}
\alias{list_osmose_versions}
\title{Returns the list of Osmose Java versions}
\usage{
list_osmose_versions()
}
\value{
List of Osmose Java versions
}
\description{
Returns the list of Osmose Java versions
}
