% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose-plot-util.R
\name{osmose.barplot}
\alias{osmose.barplot}
\title{Plots a barplot, with xlabels rotated with a 45degree angle}
\usage{
osmose.barplot(x, ...)
}
\arguments{
\item{x}{Data array}

\item{...}{Additional arguments to the barplot function (color, etc.)}
}
\description{
Plots a barplot, with xlabels rotated with a 45degree angle
}
