% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose-mortality.R
\name{osmose.extract_mort_tmean}
\alias{osmose.extract_mort_tmean}
\title{Extracts the time averaged mortality rate for a given specie.
It also eventually normalizes the mortality rates into percent.}
\usage{
osmose.extract_mort_tmean(mortMatrix, specName, repl = NULL, norm = TRUE)
}
\arguments{
\item{mortMatrix}{Mortality matrix}

\item{specName}{Name of the specie to process}

\item{repl}{Replicate index. If NULL, the replicate mean is extracted.}

\item{norm}{If TRUE, normalize the mortality rates (return percentage)}
}
\description{
Extracts the time averaged mortality rate for a given specie.
It also eventually normalizes the mortality rates into percent.
}
