% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose-diet.R
\name{osmose.extract_tmean_diet}
\alias{osmose.extract_tmean_diet}
\title{Extracts a time averaged diet matrix.
It also merges together all the prey species
for which the predation rate is less equal a given threshold.}
\usage{
osmose.extract_tmean_diet(dietMatrix, specName, repl = NULL, thres = 1)
}
\arguments{
\item{dietMatrix}{Diet matrix (data$trophic$dietMatrix)}

\item{specName}{Name of the specie.}

\item{repl}{Replicate number to extract. If NULL, the replicate averaged is extracted.}

\item{thres}{Percentage below which the predation rates will be merged}
}
\description{
Extracts a time averaged diet matrix.
It also merges together all the prey species
for which the predation rate is less equal a given threshold.
}
