% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose-diet.R
\name{osmose.extract_ts_diet}
\alias{osmose.extract_ts_diet}
\title{Extracts the time series of the diet matrix for a given 
specie. The output is also formatted into the ggplor2
stacked plot format. The predation rates below a given threshold
are merged together.}
\usage{
osmose.extract_ts_diet(dietMatrix, specName, repl = NULL, thres = 1)
}
\arguments{
\item{dietMatrix}{Diet matrix}

\item{specName}{Specie name}

\item{repl}{Replicates.}

\item{thres}{Threshold below which predation rates
are merged.}
}
\value{
A dataframe
}
\description{
Extracts the time series of the diet matrix for a given 
specie. The output is also formatted into the ggplor2
stacked plot format. The predation rates below a given threshold
are merged together.
}
