% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose.netcdf.R
\name{osmose.nc.extract_grid_mask}
\alias{osmose.nc.extract_grid_mask}
\title{Extracts the model grid file from the LTL forcing file.}
\usage{
osmose.nc.extract_grid_mask(filename, varname = "ltl_biomass",
  output_file = "ltl-grid-mask.csv", fillVal = NA)
}
\arguments{
\item{filename}{LTL forcing file name}

\item{varname}{Name of the LTL biomass variable (default "ltl_biomass")}

\item{output_file}{Name of the output file (default "ltl-grid-mask.csv")}

\item{fillVal}{Fill Value (default NA but could be 0).}
}
\description{
Extracts the model grid file from the LTL forcing file.
}
