% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose.netcdf.R
\name{osmose.nc.extract_grid_param}
\alias{osmose.nc.extract_grid_param}
\title{Computes the Osmose grid parameters by using the
LTL forcing file. The parameters are prompted in the console.
The user should copy the console output and paste it into
the Osmose configuration file.}
\usage{
extract_grid_param("ltlbiomass.nc")
}
\arguments{
\item{filename}{NetCDF filename}

\item{lonname}{Longitude name (default "longitude")}

\item{latname}{Latitude name (default "latitude")}
}
\value{
Nothing
}
\description{
Computes the Osmose grid parameters by using the
LTL forcing file. The parameters are prompted in the console.
The user should copy the console output and paste it into
the Osmose configuration file.
}
