% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose.netcdf.R
\name{osmose.nc.make_movement_netcdf}
\alias{osmose.nc.make_movement_netcdf}
\title{Converts the movement CSV settings into a set of NetCDF files 
(one per species)}
\usage{
osmose.nc.make_movement_netcdf(filename)
}
\arguments{
\item{filename}{Name of the parameter file containing the movement parameters.}
}
\description{
Converts the movement CSV settings into a set of NetCDF files 
(one per species)
}
