% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose-diet.R
\name{osmose.plot.dietMean}
\alias{osmose.plot.dietMean}
\title{Plots a time averaged diet matrix as a barplot.
It also merges together all the prey species
for which the predation rate is below a given threshold.}
\usage{
osmose.plot.dietMean(x, ...)
}
\arguments{
\item{...}{Additional arguments to the osmose.barplot function}

\item{dietMatrix}{osmose.dietMean object}
}
\description{
Plots a time averaged diet matrix as a barplot.
It also merges together all the prey species
for which the predation rate is below a given threshold.
}
