% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose-diet.R
\name{osmose.plot.dietTs}
\alias{osmose.plot.dietTs}
\title{Plot time series of osmose diet as a stacked plot}
\usage{
osmose.plot.dietTs(data, colors = NULL, ...)
}
\arguments{
\item{data}{Diet ts data}

\item{colors}{Ggplot2 color table. If NULL, a default value (scale_fill_brewer) is used.}

\item{...}{Additional arguments to the geom_area ggplot2 function}
}
\description{
Plot time series of osmose diet as a stacked plot
}
