% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose-mortality.R
\name{osmose.plot.mortMean}
\alias{osmose.plot.mortMean}
\title{Plots the time-mean mortality rates as
a barplot}
\usage{
osmose.plot.mortMean(data, ...)
}
\arguments{
\item{data}{Time-average mortality rate}

\item{...}{Additional arguments to the barplot functions}
}
\description{
Plots the time-mean mortality rates as
a barplot
}
