% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose-plot-util.R
\name{osmose.stackedpcent}
\alias{osmose.stackedpcent}
\title{Plot a stacked percent plot.}
\usage{
osmose.stackedpcent(data, legPosX = "right", legPosY = NULL, col = NULL,
  ...)
}
\arguments{
\item{data}{}

\item{legPosX}{X position of legend (if NULL, no legend is plotted)}

\item{legPosY}{Y position of legend}

\item{...}{Additional arguments to the barplot function.}
}
\description{
Plot a stacked percent plot.
}
