% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose-auxiliar.R
\name{osmose2R.v3r1}
\alias{osmose2R.v3r1}
\title{Read Osmose (version 3 release 1) outputs}
\usage{
osmose2R.v3r1(path = NULL, species.names = NULL)
}
\arguments{
\item{path}{Osmose output path}

\item{species.names}{Array of species names. If
NULL, it is extracted from file.}
}
\value{
A list of list containing the output functions
}
\description{
Read Osmose (version 3 release 1) outputs
}
