% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose-calibration.R
\name{osmose_calibration_check}
\alias{osmose_calibration_check}
\title{Check the results of a calibration trial}
\usage{
osmose_calibration_check(
  input,
  name = NULL,
  par = NULL,
  additional = NULL,
  version = "4.4.1"
)
}
\arguments{
\item{input}{Filename of the main configuration file}

\item{name}{Name for the calibration trial.}

\item{par}{A list or osmose.configuration object containing OSMOSE parameters. Only calibrated parameters are accepted.}

\item{additional}{A list or osmose.configuration object containing OSMOSE parameters. Only NON calibrated parameters are accepted, particularly output parameters.}

\item{version}{OSMOSE version. Integer (2, 3, etc.) or releases ('v3r2') are
accepted.}
}
\value{
A list of outputs from the model.
}
\description{
Check the results of a calibration trial
}
