% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose-calibration.R
\name{osmose_calibration_test}
\alias{osmose_calibration_test}
\title{Run an OSMOSE calibration test.}
\usage{
osmose_calibration_test(path, setup = NULL)
}
\arguments{
\item{path}{Path to the main calibration directory.}

\item{script}{Name of the calibratio script. The default is 'launch_calibration.R' as
automatically created by \code{\link{osmose_calibration_setup}}.}
}
\value{
TRUE is the test is completed.
}
\description{
Run an OSMOSE calibration test.
}
