% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose-templates.R
\name{osmose_template}
\alias{osmose_template}
\title{Creates OSMOSE configuration templates}
\usage{
osmose_template(input, file, what = "initialisation", control = list(), ...)
}
\arguments{
\item{input}{Filename of the main configuration file}

\item{file}{File to write the template.}

\item{what}{Type of template desired. Currently: "initialisation".}

\item{control}{Additional arguments provided for each type of template.}

\item{...}{Extra arguments passed to each method.}
}
\value{
Called by its side effects, it creates a file with the desired template.
}
\description{
Creates OSMOSE configuration templates
}
