% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose-templates.R
\name{osmose_tidy}
\alias{osmose_tidy}
\title{Create a tidy configuration compatible with OSMOSE v3.0.0 or newer}
\usage{
osmose_tidy(input, key = "reference")
}
\arguments{
\item{input}{Filename of the main configuration file}

\item{key}{Optional argument providing a name for this particular configuration.
By default, "reference" is used.}
}
\value{
Invisibily, the path to the new tidy configuration. Used mostly for the side
effect of reorganising an old configuration.
}
\description{
Create a tidy configuration compatible with OSMOSE v3.0.0 or newer
}
\examples{
\dontrun{
osmose_tidy(input=configFile)}
}
