% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose-plot-inDevel.R
\name{plot.osmose.abundanceByAge}
\alias{plot.osmose.abundanceByAge}
\title{Plots abundance by age class}
\usage{
\method{plot}{osmose.abundanceByAge}(x, species = NULL, time.mean = TRUE, ...)
}
\arguments{
\item{x}{Abundance distribution by age}

\item{species}{Species name}

\item{time.mean}{If true, time.mean abundance is plotted}

\item{...}{Additional plot arguments}
}
\description{
Plots abundance by age class
}
