% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose-plot-inDevel.R
\name{plot.osmose.abundanceBySize}
\alias{plot.osmose.abundanceBySize}
\title{Plots abundance by size class}
\usage{
\method{plot}{osmose.abundanceBySize}(x, species = NULL, time.mean = TRUE, ...)
}
\arguments{
\item{x}{Abundance distribution by size}

\item{species}{Species name}

\item{time.mean}{If true, time.mean abundance is plotted}

\item{...}{Additional plot arguments}
}
\description{
Plots abundance by size class
}
