% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose-plot-inDevel.R
\name{plot.osmose.abundanceByTL}
\alias{plot.osmose.abundanceByTL}
\title{Plots abundance by trophic level}
\usage{
\method{plot}{osmose.abundanceByTL}(x, species = NULL, time.mean = TRUE, ...)
}
\arguments{
\item{x}{Abundance distribution by age}

\item{species}{Species name}

\item{time.mean}{If true, time.mean abundance is plotted}

\item{...}{Additional plot arguments}
}
\description{
Plots abundance by trophic level
}
