% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose-plot-inDevel.R
\name{plot.osmose.biomassByAge}
\alias{plot.osmose.biomassByAge}
\title{Plots biomass by age class}
\usage{
\method{plot}{osmose.biomassByAge}(
  x,
  ts = FALSE,
  type = 1,
  species = NULL,
  speciesNames = NULL,
  start = NULL,
  end = NULL,
  initialYear = NULL,
  replicates = TRUE,
  freq = 12,
  horizontal = FALSE,
  factor = 0.001,
  xlim = NULL,
  ylim = NULL,
  col = NULL,
  border = NULL,
  lty = 1,
  lwd = 1,
  axes = TRUE,
  legend = TRUE,
  units = list(x = "years", y = "tonnes"),
  ...
)
}
\description{
Plots biomass by age class
}
