% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose-plot-inDevel.R
\name{plot.osmose.biomassBySize}
\alias{plot.osmose.biomassBySize}
\title{Plots biomass by size class}
\usage{
\method{plot}{osmose.biomassBySize}(
  x,
  ts = FALSE,
  type = 1,
  species = NULL,
  speciesNames = NULL,
  start = NULL,
  end = NULL,
  initialYear = NULL,
  replicates = TRUE,
  freq = 12,
  horizontal = FALSE,
  factor = 0.001,
  xlim = NULL,
  ylim = NULL,
  col = NULL,
  border = NULL,
  lty = 1,
  lwd = 1,
  axes = TRUE,
  legend = TRUE,
  units = list(x = "cm", y = "tonnes"),
  ...
)
}
\arguments{
\item{x}{Biomass distribution by size}

\item{species}{Species name}

\item{...}{Additional plot arguments}

\item{time.mean}{If true, time.mean biomass is plotted}
}
\description{
Plots biomass by size class
}
