% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose-plot-inDevel.R
\name{plot.osmose.biomassDistribByAge}
\alias{plot.osmose.biomassDistribByAge}
\title{Plots biomass by age class}
\usage{
\method{plot}{osmose.biomassDistribByAge}(x, species = NULL, time.mean = TRUE, ...)
}
\arguments{
\item{x}{Biomass distribution by age}

\item{species}{Species name}

\item{time.mean}{If true, time.mean biomass is plotted}

\item{...}{Additional plot arguments}
}
\description{
Plots biomass by age class
}
