% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose-plot-inDevel.R
\name{plot.osmose.biomassDistribByTL}
\alias{plot.osmose.biomassDistribByTL}
\title{Plots the biomass distribution by trophic level}
\usage{
\method{plot}{osmose.biomassDistribByTL}(x, species = NULL, time.mean = TRUE, ...)
}
\arguments{
\item{x}{Biomass by trophic level data}

\item{species}{Species name}

\item{time.mean}{If TRUE, the time-mean biomass is plotted for each TL class}

\item{...}{Additional plot arguments}
}
\description{
Plots the biomass distribution by trophic level
}
