% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose.config-class.R
\name{plot.osmose.configuration}
\alias{plot.osmose.configuration}
\title{Plot method for \code{osmose.config} objects}
\usage{
\method{plot}{osmose.configuration}(x, what = "predation", ...)
}
\arguments{
\item{x}{Object of \code{osmose.config} class. See the \code{\link{read_osmose}} 
function for more details.}

\item{what}{Variable name to plot. By default is \code{what = "predation"}. 
See Details.}

\item{...}{Additional arguments for the plot function.}
}
\description{
This method takes a \code{osmose.config} object and produce 
useful plots.
}
