% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose-plot-inDevel.R, R/osmose-plot4d.R
\name{plot.osmose.dietMatrix}
\alias{plot.osmose.dietMatrix}
\title{Plot the dietMatrix output}
\usage{
\method{plot}{osmose.dietMatrix}(
  x,
  species = NULL,
  speciesNames = NULL,
  type = 1,
  thres = 1,
  parargs = list(),
  plotargs = list(),
  legargs = list(),
  axisargs = list(),
  draw_legend = TRUE,
  norm = TRUE,
  ...
)

\method{plot}{osmose.dietMatrix}(
  x,
  species = NULL,
  speciesNames = NULL,
  type = 1,
  thres = 1,
  parargs = list(),
  plotargs = list(),
  legargs = list(),
  axisargs = list(),
  draw_legend = TRUE,
  norm = TRUE,
  ...
)
}
\arguments{
\item{x}{Object of class osmose.mortalityRate}

\item{species}{List of species to plot. If NULL, all species are drawn.}

\item{thres}{Threshold (in percentage) below which preys are concatenated.}

\item{norm}{True if percentage is returned instead of raw mort. rates}

\item{...}{Additional parameters to the barplot/lines functions}

\item{add_text}{True if values sjopu}

\item{color}{Histogram colors (if NULL, gray). Only used if time.mean=T}

\item{time.mean}{True if time-average plot should be ok}
}
\description{
Plot the dietMatrix output

Plot mortality rates
}
