% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose-plot-inDevel.R
\name{plot.osmose.dietMatrixbySize}
\alias{plot.osmose.dietMatrixbySize}
\title{Plot diet matrix by size class}
\usage{
\method{plot}{osmose.dietMatrixbySize}(
  x,
  species = NULL,
  speciesNames = NULL,
  norm = TRUE,
  type = 1,
  parargs = list(),
  plotargs = list(),
  legargs = list(),
  axisargs = list(),
  draw_legend = TRUE,
  ...
)
}
\arguments{
\item{x}{Data list}

\item{species}{List of species name. If null, all species are plotted}

\item{...}{Extra arguments of the graphics barplot function}
}
\description{
Plot diet matrix by size class
}
