% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose-plot-inDevel.R
\name{plot.osmose.meanTLByAge}
\alias{plot.osmose.meanTLByAge}
\title{Plots mean trophic level by age class}
\usage{
\method{plot}{osmose.meanTLByAge}(x, species = NULL, time.mean = TRUE, ...)
}
\arguments{
\item{x}{Trophic level distribution by age}

\item{species}{Species name}

\item{time.mean}{If true, time.mean TL is plotted}

\item{...}{Extra arguments passed to the method.}
}
\description{
Plots mean trophic level by age class
}
