% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose-plot-inDevel.R
\name{plot.osmose.meanTLBySize}
\alias{plot.osmose.meanTLBySize}
\title{Plots mean trophic level by size class}
\usage{
\method{plot}{osmose.meanTLBySize}(x, species = NULL, time.mean = TRUE, ...)
}
\arguments{
\item{x}{TL distribution by size}

\item{species}{Species name}

\item{time.mean}{If true, time.mean TL is plotted}

\item{...}{Extra arguments passed to the method.}
}
\description{
Plots mean trophic level by size class
}
