% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose-plot4d.R
\name{plot.osmose.mortalityRate}
\alias{plot.osmose.mortalityRate}
\title{Plot mortality rates}
\usage{
\method{plot}{osmose.mortalityRate}(
  x,
  species = NULL,
  speciesNames = NULL,
  norm = TRUE,
  type = 1,
  draw_legend = TRUE,
  plotargs = list(),
  parargs = list(),
  legargs = list(),
  axisargs = list(),
  ...
)
}
\arguments{
\item{x}{Object of class osmose.mortalityRate}

\item{species}{List of species to plot. If NULL, all species are drawn.}

\item{norm}{True if percentage is returned instead of raw mort. rates}

\item{type}{1 if stacked barplot, 2 if pie chart}

\item{...}{Additional parameters to the barplot/lines functions}

\item{time.mean}{True if time-average plot should be ok}
}
\description{
Plot mortality rates
}
