% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose-plot-inDevel.R
\name{plot.osmose.mortalityRateByAge}
\alias{plot.osmose.mortalityRateByAge}
\title{Plots mortality rates by age class.}
\usage{
\method{plot}{osmose.mortalityRateByAge}(
  x,
  species = NULL,
  speciesNames = NULL,
  norm = TRUE,
  type = 1,
  parargs = list(),
  plotargs = list(),
  legargs = list(),
  axisargs = list(),
  draw_legend = TRUE,
  ...
)
}
\arguments{
\item{x}{Mortality object}

\item{species}{Species name}

\item{norm}{Wheter mortality should be normalized (only if time.mean=TRUE)}

\item{...}{Additional arguments of the function.}

\item{time.mean}{If time-average should be plotted. If TRUE, mortality rates by
for each age class and for each mortality types are plotted. Else, the time series for
a specific type of mortality is plotted.}

\item{mtype}{Mortality type ("Mtot", Mpred", "Mstar", "Mnat", "F" or "Z")}
}
\description{
Plots mortality rates by age class.
}
