% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose-plot.R
\name{plot.osmose.mortalityRateDistribBySize}
\alias{plot.osmose.mortalityRateDistribBySize}
\title{Plots mortality rates by size class.}
\usage{
\method{plot}{osmose.mortalityRateDistribBySize}(x, species = NULL, time.mean = TRUE, norm = TRUE, mtype = "Mtot", ...)
}
\arguments{
\item{x}{Mortality object}

\item{species}{Species name}

\item{time.mean}{If time-average should be plotted. If TRUE, mortality rates by
for each age class and for each mortality types are plotted. Else, the time series for
a specific type of mortality is plotted.}

\item{norm}{Wheter mortality should be normalized (only if time.mean=TRUE)}

\item{mtype}{Mortality type ("Mtot", Mpred", "Mstar", "Mnat", "F" or "Z")}

\item{...}{Additional arguments of the function.}
}
\value{
None
}
\description{
Plots mortality rates by size class.
}
