% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose-plot.R
\name{plot.osmose.output.abundance}
\alias{plot.osmose.output.abundance}
\title{Plots abundance time-series}
\usage{
\method{plot}{osmose.output.abundance}(x, start = NULL, conf = 0.95,
  factor = 1e-06, replicates = FALSE, freq = 12, alpha = 0.5,
  col = "black", xlim = NULL, ylim = NULL, nrep = NULL,
  time.mean = FALSE, species = NULL, nmax = NULL, ...)
}
\arguments{
\item{x}{Abundance data}

\item{start}{First simulation year}

\item{conf}{Confidence interval}

\item{factor}{Multiplication factor}

\item{replicates}{Draws the time-series for each replicate}

\item{freq}{Output frequency (used to convert from time-step into years)}

\item{alpha}{Transparency for confidence interval}

\item{col}{Line color}

\item{xlim}{Xaxis limit}

\item{ylim}{Yaxis limit}

\item{nrep}{Maximum number of replicate to draw. If NULL, all the replicates
are drawn. Only used if replicates=TRUE}

\item{time.mean}{If FALSE, the mean biomass is drawn for each specie as a barplot.}

\item{species}{Name of the species to draw (only if time.mean=FALSE)}

\item{nmax}{Maximum number of species to draw (only used if time.mean=TRUE)}

\item{...}{}
}
\description{
Plots abundance time-series
}
