% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose-plot-nico.R
\name{plot.osmose.output.biomassDistribBySize}
\alias{plot.osmose.output.biomassDistribBySize}
\title{Plots biomass by size class}
\usage{
\method{plot}{osmose.output.biomassDistribBySize}(data, species = NULL,
  time.mean = FALSE, lwd = 2, ...)
}
\arguments{
\item{data}{Biomass distribution by size}

\item{species}{Species name}

\item{time.mean}{If true, time.mean biomass is plotted}

\item{lwd}{Line width}

\item{...}{}
}
\description{
Plots biomass by size class
}
