% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose-plot-nico.R
\name{plot.osmose.output.biomassDistribByTL}
\alias{plot.osmose.output.biomassDistribByTL}
\title{Plots the biomass distribution by trophic level}
\usage{
\method{plot}{osmose.output.biomassDistribByTL}(data, species = NULL,
  time.mean = TRUE, lwd = 2, thres = 1, colors = NULL, ...)
}
\arguments{
\item{data}{Biomass by trophic level data}

\item{species}{Species name}

\item{time.mean}{If TRUE, the time-mean biomass is plotted for each TL class}

\item{lwd}{Line width}

\item{thres}{Threshold below which data are not plotted. If NULL, all data are plotted. Only if time.mean=TRUE}

\item{colors}{GGplot2 colors}

\item{...}{}
}
\description{
Plots the biomass distribution by trophic level
}
