% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose-plot-nico.R
\name{plot.osmose.output.dietMatrix}
\alias{plot.osmose.output.dietMatrix}
\title{Plots diet matrix}
\usage{
\method{plot}{osmose.output.dietMatrix}(x, time.mean = FALSE,
  species = NULL, colors = NULL, thres = 1, ...)
}
\arguments{
\item{x}{Diet matrix}

\item{time.mean}{If TRUE, the time mean diet matrix is computed is displayed.}

\item{species}{Species name}

\item{colors}{GGplot2 colors (see for instance scale_fill_manual)}

\item{thres}{Thresholds (in percentage). Time-average predation rates below this threashold 
are binned together ("other" column).}

\item{...}{}
}
\value{
None
}
\description{
Plots diet matrix
}
