% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose-plot-nico.R
\name{plot.osmose.output.mortalityRate}
\alias{plot.osmose.output.mortalityRate}
\title{Plots mortality rates}
\usage{
\method{plot}{osmose.output.mortalityRate}(data, species = NULL,
  time.mean = TRUE, stade = NULL, norm = TRUE, colors = NULL, ...)
}
\arguments{
\item{data}{Mortality rate}

\item{species}{Species name}

\item{time.mean}{If TRUE, the mortality rates for each life stade and for each mortality
types are plotted. If FALSE, the time series for a given life stade are plotted.}

\item{stade}{Stade ("eggs", "juveniles", "adults). Only if time.mean=FALSE}

\item{norm}{Whether percentage instead of raw mortality rates should be plotted. 
Only if time.mean=TRUE}

\item{colors}{Ggplot2 color array (see for instance scale_fill_manual)}

\item{...}{}
}
\description{
Plots mortality rates
}
