% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose-plot-nico.R
\name{plot.osmose.output.predatorPressure}
\alias{plot.osmose.output.predatorPressure}
\title{Plots predator pressure.}
\usage{
\method{plot}{osmose.output.predatorPressure}(data, time.mean = FALSE,
  species = NULL, colors = NULL, nmax = NULL, ...)
}
\arguments{
\item{data}{Predator pressure.}

\item{time.mean}{If TRUE, plots the time-average predator pressure 
sorted in decreasing order}

\item{species}{Species names}

\item{colors}{GGplot2 colors}

\item{nmax}{Maximum number of values to draw. If NULL, all data are plotted.}

\item{...}{}
}
\description{
Plots predator pressure.
}
