% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose-plot-inDevel.R
\name{plot.osmose.predatorPressure}
\alias{plot.osmose.predatorPressure}
\title{Plots predator pressure.}
\usage{
\method{plot}{osmose.predatorPressure}(x, species = NULL, time.mean = TRUE, ...)
}
\arguments{
\item{x}{Predator pressure.}

\item{species}{Species names}

\item{time.mean}{If TRUE, plots the time-average predator pressure 
sorted in decreasing order}

\item{...}{Additional plot arguments}
}
\description{
Plots predator pressure.
}
