% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose-class.R, R/osmose-plot.R
\name{plot.osmose}
\alias{plot.osmose}
\alias{plot.osmose.biomass}
\alias{plot.osmose.abundance}
\alias{plot.osmose.yield}
\alias{plot.osmose.yieldByYear}
\alias{plot.osmose.yieldN}
\alias{plot.osmose.biomassByTL}
\alias{plot.osmose.biomassBySize}
\alias{plot.osmose.biomassByAge}
\alias{plot.osmose.abundanceByTL}
\alias{plot.osmose.abundanceBySize}
\alias{plot.osmose.abundanceByAge}
\alias{plot.osmose.yieldNBySize}
\alias{plot.osmose.yieldNByAge}
\alias{plot.osmose.yieldBySize}
\alias{plot.osmose.yieldByAge}
\title{Plot method for osmose objects}
\usage{
\method{plot}{osmose}(x, what = "biomass", ...)

\method{plot}{osmose.biomass}(
  x,
  ts = TRUE,
  type = 1,
  species = NULL,
  speciesNames = NULL,
  start = NULL,
  end = NULL,
  initialYear = NULL,
  replicates = TRUE,
  freq = 12,
  horizontal = FALSE,
  conf = 0.95,
  factor = 0.001,
  xlim = NULL,
  ylim = NULL,
  col = NULL,
  alpha = NULL,
  border = NULL,
  lty = 1,
  lwd = 1,
  axes = TRUE,
  legend = TRUE,
  units = "tonnes",
  zero = FALSE,
  ...
)

\method{plot}{osmose.abundance}(
  x,
  species = NULL,
  speciesNames = NULL,
  start = NULL,
  end = NULL,
  initialYear = NULL,
  ts = TRUE,
  type = 1,
  replicates = TRUE,
  freq = 12,
  horizontal = FALSE,
  conf = 0.95,
  factor = 0.001,
  xlim = NULL,
  ylim = NULL,
  col = NULL,
  alpha = NULL,
  border = NULL,
  lty = 1,
  lwd = 1,
  axes = TRUE,
  legend = TRUE,
  units = "individuals",
  ...
)

\method{plot}{osmose.yield}(
  x,
  species = NULL,
  speciesNames = NULL,
  start = NULL,
  end = NULL,
  initialYear = NULL,
  ts = TRUE,
  type = 1,
  replicates = TRUE,
  freq = 12,
  horizontal = FALSE,
  conf = 0.95,
  factor = 0.001,
  xlim = NULL,
  ylim = NULL,
  col = NULL,
  alpha = NULL,
  border = NULL,
  lty = 1,
  lwd = 1,
  axes = TRUE,
  legend = TRUE,
  units = "tonnes",
  ...
)

\method{plot}{osmose.yieldByYear}(
  x,
  species = NULL,
  speciesNames = NULL,
  start = NULL,
  end = NULL,
  initialYear = NULL,
  ts = TRUE,
  type = 1,
  replicates = TRUE,
  freq = 12,
  horizontal = FALSE,
  conf = 0.95,
  factor = 0.001,
  xlim = NULL,
  ylim = NULL,
  col = NULL,
  alpha = NULL,
  border = NULL,
  lty = 1,
  lwd = 2,
  axes = TRUE,
  legend = TRUE,
  units = "tonnes",
  ...
)

\method{plot}{osmose.yieldN}(
  x,
  species = NULL,
  speciesNames = NULL,
  start = NULL,
  end = NULL,
  initialYear = NULL,
  ts = TRUE,
  type = 1,
  replicates = TRUE,
  freq = 12,
  horizontal = FALSE,
  conf = 0.95,
  factor = 0.001,
  xlim = NULL,
  ylim = NULL,
  col = NULL,
  alpha = NULL,
  border = NULL,
  lty = 1,
  lwd = 1,
  axes = TRUE,
  legend = TRUE,
  units = "individuals",
  ...
)

\method{plot}{osmose.biomassByTL}(
  x,
  type = 1,
  species = NULL,
  speciesNames = NULL,
  start = NULL,
  end = NULL,
  freq = 12,
  horizontal = FALSE,
  factor = 0.001,
  xlim = NULL,
  ylim = NULL,
  col = NULL,
  border = NULL,
  legend = TRUE,
  lty = 1,
  lwd = 1,
  axes = TRUE,
  units = list(x = "", y = "tonnes"),
  ...
)

\method{plot}{osmose.biomassBySize}(
  x,
  type = 1,
  species = NULL,
  speciesNames = NULL,
  start = NULL,
  end = NULL,
  freq = 12,
  horizontal = FALSE,
  factor = 0.001,
  xlim = NULL,
  ylim = NULL,
  col = NULL,
  border = NULL,
  legend = TRUE,
  lty = 1,
  lwd = 1,
  axes = TRUE,
  units = list(x = "cm", y = "tonnes"),
  ...
)

\method{plot}{osmose.biomassByAge}(
  x,
  type = 1,
  species = NULL,
  speciesNames = NULL,
  start = NULL,
  end = NULL,
  freq = 12,
  horizontal = FALSE,
  factor = 0.001,
  xlim = NULL,
  ylim = NULL,
  col = NULL,
  border = NULL,
  legend = TRUE,
  lty = 1,
  lwd = 1,
  axes = TRUE,
  units = list(x = "years", y = "tonnes"),
  ...
)

\method{plot}{osmose.abundanceByTL}(
  x,
  type = 1,
  species = NULL,
  speciesNames = NULL,
  start = NULL,
  end = NULL,
  freq = 12,
  horizontal = FALSE,
  factor = 0.001,
  xlim = NULL,
  ylim = NULL,
  col = NULL,
  border = NULL,
  legend = TRUE,
  lty = 1,
  lwd = 1,
  axes = TRUE,
  units = list(x = "", y = "individuals"),
  ...
)

\method{plot}{osmose.abundanceBySize}(
  x,
  type = 1,
  species = NULL,
  speciesNames = NULL,
  start = NULL,
  end = NULL,
  freq = 12,
  horizontal = FALSE,
  factor = 0.001,
  xlim = NULL,
  ylim = NULL,
  col = NULL,
  border = NULL,
  legend = TRUE,
  lty = 1,
  lwd = 1,
  axes = TRUE,
  units = list(x = "cm", y = "individuals"),
  ...
)

\method{plot}{osmose.abundanceByAge}(
  x,
  type = 1,
  species = NULL,
  speciesNames = NULL,
  start = NULL,
  end = NULL,
  freq = 12,
  horizontal = FALSE,
  factor = 0.001,
  xlim = NULL,
  ylim = NULL,
  col = NULL,
  border = NULL,
  legend = TRUE,
  lty = 1,
  lwd = 1,
  axes = TRUE,
  units = list(x = "years", y = "individuals"),
  ...
)

\method{plot}{osmose.yieldNBySize}(
  x,
  type = 1,
  species = NULL,
  speciesNames = NULL,
  start = NULL,
  end = NULL,
  freq = 12,
  horizontal = FALSE,
  factor = 0.001,
  xlim = NULL,
  ylim = NULL,
  col = NULL,
  border = NULL,
  legend = TRUE,
  lty = 1,
  lwd = 1,
  axes = TRUE,
  units = list(x = "cm", y = "individuals"),
  ...
)

\method{plot}{osmose.yieldNByAge}(
  x,
  type = 1,
  species = NULL,
  speciesNames = NULL,
  start = NULL,
  end = NULL,
  freq = 12,
  horizontal = FALSE,
  factor = 0.001,
  xlim = NULL,
  ylim = NULL,
  col = NULL,
  border = NULL,
  legend = TRUE,
  lty = 1,
  lwd = 1,
  axes = TRUE,
  units = list(x = "years", y = "individuals"),
  ...
)

\method{plot}{osmose.yieldBySize}(
  x,
  type = 1,
  species = NULL,
  speciesNames = NULL,
  start = NULL,
  end = NULL,
  freq = 12,
  horizontal = FALSE,
  factor = 0.001,
  xlim = NULL,
  ylim = NULL,
  col = NULL,
  border = NULL,
  legend = TRUE,
  lty = 1,
  lwd = 1,
  axes = TRUE,
  units = list(x = "cm", y = "tonnes"),
  ...
)

\method{plot}{osmose.yieldByAge}(
  x,
  type = 1,
  species = NULL,
  speciesNames = NULL,
  start = NULL,
  end = NULL,
  freq = 12,
  horizontal = FALSE,
  factor = 0.001,
  xlim = NULL,
  ylim = NULL,
  col = NULL,
  border = NULL,
  legend = TRUE,
  lty = 1,
  lwd = 1,
  axes = TRUE,
  units = list(x = "years", y = "tonnes"),
  ...
)
}
\arguments{
\item{x}{\code{osmose} object.}

\item{what}{Variable name to plot. By default is \code{what = "biomass"}. 
See Details}

\item{...}{Extra arguments of the function. See Details.}

\item{ts}{\code{logical} parameter. By default \code{ts = TRUE} and the plots 
are going to be time series where the x-axis is the time. For plots where the 
x-axis is not the time \code{ts = FALSE}.}

\item{type}{A \code{numeric} value, indicating the type of plot to be used 
(\code{type = 1}, by default). See Details.}

\item{species}{A \code{numeric} vector specifying the species that will be 
showed. If \code{NULL} (default), all the available species time series will 
be plotted. See Details.}

\item{speciesNames}{A vector with the names for each species. If \code{NULL}
(default) the names of \code{x} will be used.}

\item{start}{A \code{numeric} value indicating the first element of the time 
indexation of the object \code{x} is specified with this parameter. By 
default \code{start = NULL} and the time indexation start with the first 
element of \code{x}.}

\item{end}{A \code{numeric} value. The last element of the time indexation of
the object \code{x} is specified with this parameter. By default 
\code{end = NULL} and the time indexation finish with the last element of 
\code{x}.}

\item{initialYear}{A \code{numeric} value. It specifies the first element 
that is going to be used on the x axis for the plots.}

\item{replicates}{\code{logical}. It controls to show whether the values in 
each simulation (\code{TRUE}) or the median of the values along the time 
(\code{FALSE}).}

\item{freq}{A \code{numeric} value to indicate the steps by year used in the 
time series. Default values, see Details.}

\item{horizontal}{a \code{logical} value. If \code{FALSE}, plot types that do 
not represent time series (boxplots and barplots) will be drawn horizontally.}

\item{conf}{\code{numeric} single value indicating the confidence interval 
showed in the plot as shadow. By default \code{conf = 0.95}.}

\item{factor}{A number to indicate the scale of the variable on the y-axis. 
By default \code{factor = 1e-3}.}

\item{xlim, ylim}{numeric vectors of length 2, giving the x and y coordinates 
ranges.}

\item{col}{A vector with the color names for the plots. By default 
\code{col = NULL} and the colors will be chosen by the function.}

\item{alpha}{factor modifying the opacity alpha given to confidence interval 
(CI) polygons (check \code{\link{adjustcolor}}). By default (\code{NULL}), 
its value will depend on the plot type. See Details.}

\item{border}{the color to draw the border of CI polygons, bar plots and 
box plots. The default value (\code{NULL}) will depend on the type, see Details.}

\item{lty}{The line type (check \code{\link{par}}).}

\item{lwd}{The line width (check \code{\link{par}}).}

\item{axes}{a \code{logical} value indicating whether both axes should be 
drawn on the plot.}

\item{legend}{\code{logical} do you want to show a legend of species? (valid
only for those plot types TS-2)}

\item{units}{String with units of the variable that defines the label set 
on top of the plot. If \code{NULL} (default), the value will depend on the
class, see Details.}
}
\value{
A graph of an \code{osmose} object.
}
\description{
This method takes a \code{osmose} object and produce useful 
plots

This function implements a plot method for different osmose 
classes.
}
\details{
\code{what} argument can take next values:
\itemize{
 \item "biomass" to plot the species biomass (see \link{plot.osmose.biomass})
 \item "abundance" to plot the species abundance (see \link{plot.osmose.abundance})
 \item "yield" to plot the species yield (see \link{plot.osmose.yield})
 \item "yieldN" to plot the species yield (in numbers) (see \link{plot.osmose.yieldN})
}

User can select within different plot types depending on the osmose
class. So, for classes \code{biomass}, \code{abundance}, \code{yield} and 
\code{yieldN} (\strong{Group 1}) the user must use \code{ts} and \code{type} 
arguments to specify the type of plot to draw.

So, for the \strong{Group 1} and \code{ts = TRUE}, user will select within:
\itemize{
 \item{\code{type = 1}: }{Generates a matrix plot of the selected variable 
 plotted independently.}
 \item{\code{type = 2}: }{Generates a single plot with overlaped time series.
 You can include confidence interval shadow around lines.}
 \item{\code{type = 3}: }{Generates a single plot with cumulated time series.
 So, the method will sort and overlap the time series values for the selected 
 variable}
 \item{\code{type = 4}: }{ONLY valid for a single species. It generates a bar 
 plot of the time series.}
}

For the Group 1 and \code{ts = FALSE}, user will select within:
\itemize{
 \item{\code{type = 1}: }{Generates a bar plot of the selected variable by 
 species, including interval confidence bars.}
 \item{\code{type = 2}: }{Generates a boxplot of the selected variable by 
 species.}
}

The \strong{Group 2} refers classes of \strong{Group 1} with \strong{By} 
specification (e.g. \code{yieldNBySize}, \code{abundanceByAge},
\code{biomassByTL}). There are 2 available plot types:

\itemize{
 \item{\code{type = 1}: }{Generates a bar plot of the selected variable by 
 species. The \code{...} will be passed to the internal \link{barplot} 
 function.}
 \item{\code{type = 2}: }{Generates a plot using \link{image.plot} function,
 placing the species in the left axis, the By-variable (e.g. Size) on the 
 bottom and the main variable (e.g. biomass) in a color scale. The \code{...}
 will be passed to the internal \link{image.plot} function.}
}

Both \code{alpha} and \code{border} controls the opacity and the color of 
border, respectively, for polygons, bar plots and box plots. So depending on 
the \code{ts}-\code{type}, they will have different default values: 
\itemize{
 \item{\code{alpha = 0.3} & \code{border = NA} for }{\code{TS = TRUE} with 
 \code{type = 1 or 2}.}
 \item{\code{alpha = 1} & \code{border = TRUE} for }{Everything else.}
}

\code{species} argument follows the indexation way of java: starting in zero,
as osmose-java returns species outputs.

Default value for \code{freq} will be calculated from \code{x}: 
\eqn{freq = 1/x$model$start}.

\code{units} could be whether a single character vector (currently applied on
classes \strong{without By}, e.g. \code{osmose.biomass}, 
\code{osmose.yieldN}) or a list of length 2 specifying the units for x and y 
(currently used with classes \strong{with By}, e.g. 
\code{plot.osmose.biomassBySize}, \code{plot.osmose.yieldNByAge}).

Extra arguments can be passed from \link{plot.default} using \code{...}: 
\code{cex}, \code{cex.axis}, \code{border} (useful for \link{polygon}, 
\link{boxplot}s and \link{barplot}s), etc. It is important to notice that,
depending on where the ellipsys is passed, it allows and refuse the use of
certain arguments. For instance, if the selected plot is a bar plot, the 
\code{...} will be passed to the internal \code{barplot} call, so if 
\code{cex} argument is used, it will create a conflict and an error (a 
typical error of using \code{cex} in \link{barplot} function).
}
\author{
Criscely Lujan Paredes
}
