% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose-plot.R
\name{plot.osmose.yieldNByAge}
\alias{plot.osmose.yieldNByAge}
\title{Plots yield abundance by age class}
\usage{
\method{plot}{osmose.yieldNByAge}(x, species = NULL, time.mean = TRUE, ...)
}
\arguments{
\item{x}{Biomass distribution by size}

\item{species}{Species name}

\item{time.mean}{If true, time.mean biomass is plotted}

\item{...}{Extra arguments passed to the method.}
}
\description{
Plots yield abundance by age class
}
