% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose-class.R
\name{print.osmose}
\alias{print.osmose}
\title{Print information for an \code{osmose} object}
\usage{
\method{print}{osmose}(x, ...)
}
\arguments{
\item{x}{\code{osmose} class object (see the \code{\link{read_osmose}} 
function).}

\item{...}{Additional arguments for \code{print}.}
}
\description{
Print information for an \code{osmose} object
}
