% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose-class.R
\name{print.summary.osmose}
\alias{print.summary.osmose}
\title{Print the summary informations about Osmose outputs}
\usage{
\method{print}{summary.osmose}(x, ...)
}
\arguments{
\item{x}{Osmose outputs (see the \code{\link{osmose2R}} function)}
}
\description{
Print the summary informations about Osmose outputs
}
