% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose-internal.R
\name{readOsmoseConfiguration}
\alias{readOsmoseConfiguration}
\title{Reads Osmose configuration files.}
\usage{
readOsmoseConfiguration(file, config = NULL, absolute = TRUE)
}
\arguments{
\item{file}{Main configuration file}

\item{config}{Configuration object to which file parameters are appended}

\item{absolute}{Whether the path is absolute (TRUE) or relative (FALSE)}
}
\value{
A list tree.
}
\description{
Reads Osmose configuration files.
}
