% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose-auxiliar.R
\name{readOsmoseFiles}
\alias{readOsmoseFiles}
\title{Read Osmose output file}
\usage{
readOsmoseFiles(
  path,
  type,
  bySpecies = FALSE,
  bySurvey = FALSE,
  ext = "csv",
  varid = NA,
  ...
)
}
\arguments{
\item{path}{Osmose output path}

\item{type}{Data type ("biomass", etc)}

\item{bySpecies}{TRUE if should read one file per species.}

\item{ext}{The extension of the files shich will be read.}

\item{...}{Additional arguments}
}
\value{
Output data frame
}
\description{
Read Osmose output file
}
