% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose-getVar.R
\name{summary.osmose.dietMatrix}
\alias{summary.osmose.dietMatrix}
\title{Title}
\usage{
\method{summary}{osmose.dietMatrix}(object, species = NULL, thres = 1, ...)
}
\arguments{
\item{object}{an object of class \code{osmose.mortalityRate} for which a summary is desired.}

\item{species}{Name of the species to get a summary.}

\item{thres}{Threshold which is used to keep values of species matrix.}

\item{...}{Extra arguments passed to the method.}
}
\description{
Title
}
