% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose-class.R
\name{summary.osmose}
\alias{summary.osmose}
\title{Summarizes informations about Osmose outputs}
\usage{
\method{summary}{osmose}(object, ...)
}
\arguments{
\item{object}{Osmose outputs (see the \code{\link{read_osmose}} function)}

\item{...}{Additional arguments for \code{summary}.}
}
\description{
Summarizes informations about Osmose outputs
}
