% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_config.R
\name{update_ltl}
\alias{update_ltl}
\title{Update LTL ncdf files to version 4 format}
\usage{
update_ltl(input, filename = NULL, absolute = TRUE)
}
\arguments{
\item{input}{Filename of the main configuration file}

\item{filename}{Filename of the NetCDF LTL file}

\item{absolute}{Whether the input file uses absolute (TRUE) or relative (FALSE) paths.}
}
\description{
Update LTL ncdf files to version 4 format
}
