% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose_init-update_maps.R
\name{update_maps}
\alias{update_maps}
\title{Update OSMOSE maps from csv (v3.x) to ncdf (v4.2+)}
\usage{
update_maps(input, output, conf, sep = ",", na.strings = -99, test = FALSE)
}
\arguments{
\item{input}{Path to the input configuration file (v3)}

\item{output}{Path to the folder to store the ncdf maps}

\item{conf}{Path to the output configuration file (v4)}

\item{sep}{Separator used in the csv maps, default to comma.}

\item{na.strings}{Value used for 'land' in the maps, default to -99.}
}
\value{
The side effect is to create ncdf files with the maps.
}
\description{
Update OSMOSE maps from csv (v3.x) to ncdf (v4.2+)
}
