% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_config.R
\name{update_osmose}
\alias{update_osmose}
\title{Update OSMOSE configuration}
\usage{
update_osmose(
  input,
  log = "osmose.log",
  version = "3.3.3",
  osmose = NULL,
  java = "java",
  verbose = TRUE,
  absolute = TRUE
)
}
\arguments{
\item{input}{path to the main configuration file}

\item{log}{File to save OSMOSE execution messages.}

\item{version}{target OSMOSE version for the update.}

\item{osmose}{Path to a OSMOSE .jar executable. By default (NULL), uses the stable 
jar for the current version.}

\item{java}{Path to the java executable. The default assumes 'java' is 
on the search path.}

\item{verbose}{Show messages? (output in the log file if FALSE).}

\item{absolute}{Whether the input file uses absolute (\code{TRUE}, default) or relative (\code{FALSE}) paths.}
}
\description{
Update OSMOSE configuration
}
