% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacheManager.R
\name{updateCache}
\alias{updateCache}
\title{Update the cache for the different function}
\usage{
updateCache(nameFile)
}
\arguments{
\item{nameFile}{The name of the file you want to update. If it's `NULL` the function will update all the files.}
}
\description{
This function update the cache for the environmental variables:
   - wc2-5
   - CWD
   - E
}
\examples{
\dontrun{
updateCache()
}
}
\author{
Arthur PERE
}
