% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose.interannual.R
\name{writeFishingFiles}
\alias{writeFishingFiles}
\title{Write fishing files from osmose configuration list.}
\usage{
writeFishingFiles(L1, outputPath)
}
\arguments{
\item{L1}{Osmose configuration (see \code{\link{readOsmoseConfiguration}})}

\item{outputPath}{Output path}
}
\description{
Write fishing files from osmose configuration list.
}
