% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose.maps.R
\name{writeMaps}
\alias{writeMaps}
\title{Write maps for osmose species distribution}
\usage{
writeMaps(object, files = NULL, outdir = "maps")
}
\arguments{
\item{object}{}

\item{files}{}

\item{outdir}{}
}
\description{
Write maps for osmose species distribution
}

