% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacheManager.R
\name{cacheManager}
\alias{cacheManager}
\title{Manage the cache}
\usage{
cacheManager(nameFile)
}
\arguments{
\item{nameFile}{the name of the file or folder}
}
\value{
the path to the file we need
}
\description{
The function creates a folder (only once) and then controls files that need to be downloaded and
placed in this folder.
}
\section{Localisation}{

The localisation of the folder is defined in the `.Renviron` file, by setting the `OSMOSE_DIR`
environment variable:
  - On Linux/Mac Os X : `OSMOSE_DIR=/Users/Nicolas/Desktop/OSMOSE_TEST/R`
  - On Windows: `OSMOSE_DIR=C:\\Users\\Nicolas\\Desktop\\OSMOSE_TEST\\R`

If this variable is not set, the files will be downloaded into a temporary directory.
}

\author{
Arthur PERE

Nicolas BARRIER
}
\keyword{Internal}
