% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose-calibration-auxiliar.R
\name{colsum}
\alias{colsum}
\alias{colsum.default}
\alias{colsum.matrix}
\alias{colsum.array}
\title{Give Row or Colums Sums of a Matrix or Array, Based on a Grouping Variable}
\usage{
colsum(x, group, reorder = TRUE, ...)

\method{colsum}{default}(x, group, reorder = TRUE, na.rm = FALSE, ...)

\method{colsum}{matrix}(x, group, reorder = TRUE, na.rm = FALSE, ...)

\method{colsum}{array}(x, group, reorder = TRUE, na.rm = FALSE, ...)
}
\arguments{
\item{x}{a matrix, data frame or vector of numeric data.  Missing
    values are allowed.  A numeric vector will be treated as a column vector.}

\item{group}{a vector or factor giving the grouping, with one element per column of x. Missing values will be treated as another group and a warning will be given.}

\item{reorder}{if \code{TRUE}, then the result will be in order of
    \code{sort(unique(group))}, if \code{FALSE}, it will be in the order
    that groups were encountered.}

\item{...}{other arguments to be passed to or from methods.}

\item{na.rm}{logical (\code{TRUE} or \code{FALSE}).  Should \code{NA}
    (including \code{NaN}) values be discarded?}
}
\description{
Give Row or Colums Sums of a Matrix or Array, Based on a Grouping Variable
}
