% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose.maps.R
\name{createOsmoseMaps}
\alias{createOsmoseMaps}
\title{Create osmose maps}
\usage{
createOsmoseMaps(..., outdir = NULL, confdir = NULL, write = TRUE,
  confile = "maps-parameters.csv", prefixPath = "maps")
}
\arguments{
\item{...}{Several object containing maps for species distribution}

\item{outdir}{}

\item{confdir}{}

\item{write}{Boolean, write maps to csv files for osmose?}

\item{confile}{File to save osmose configuration for choosen maps.}
}
\description{
Create osmose maps
}
