% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose-internal.R
\name{existOsmoseParameter}
\alias{existOsmoseParameter}
\title{Check if a parameter exists}
\usage{
existOsmoseParameter(par, ..., keep.att = FALSE)
}
\arguments{
\item{par}{Output of the \code{link{read_osmose}} function}

\item{...}{String arguments}

\item{keep.att}{Whether parameter attributes should be kept}
}
\description{
Check if a parameter exists
}
