% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose.netcdf.R
\name{extract_grid_mask}
\alias{extract_grid_mask}
\title{Extracts the model grid file from the LTL forcing file.}
\usage{
extract_grid_mask(filename, varname = "ltl_biomass",
  output_file = "ltl-grid-mask.csv", fillVal = NA, ...)
}
\arguments{
\item{filename}{LTL forcing file name}

\item{varname}{Name of the LTL biomass variable (default "ltl_biomass")}

\item{output_file}{Name of the output file (default "ltl-grid-mask.csv")}

\item{fillVal}{Fill Value (default NA but could be 0).}

\item{...}{Additional arguments of the ncvar_get function. Used to allow more flexibility on the choice of the
mask variable via the start and count arguments. If your mask variable is already 2D, nothing to do. If it is 3D, you
need to set the start/count arrays so that the final variable is of dimensions (lon, lat).}
}
\description{
Extracts the model grid file from the LTL forcing file.
}
