% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose.netcdf.R
\name{extract_grid_param}
\alias{extract_grid_param}
\title{Computes the Osmose grid parameters by using the
LTL forcing file. The parameters are prompted in the console.
The user should copy the console output and paste it into
the Osmose configuration file.}
\usage{
extract_grid_param("ltlbiomass.nc")
}
\arguments{
\item{filename}{NetCDF filename}

\item{lonname}{Longitude name (default "longitude")}

\item{latname}{Latitude name (default "latitude")}

\item{csv}{Output CSV file (if NULL, no file is written). The directory is created recursively
if it does not exist}
}
\value{
The grid format as a data frame.
}
\description{
Computes the Osmose grid parameters by using the
LTL forcing file. The parameters are prompted in the console.
The user should copy the console output and paste it into
the Osmose configuration file.
}
