% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose-readConfiguration.R
\name{get_par}
\alias{get_par}
\alias{get_species}
\alias{get_fisheries}
\alias{get_surveys}
\title{Get a parameter from a configuration file.}
\usage{
get_par(
  conf,
  par = NULL,
  sp = NULL,
  fsh = NULL,
  sr = NULL,
  invert = FALSE,
  as.is = FALSE,
  unlist = FALSE,
  linear = FALSE
)

get_species(
  x,
  type = NULL,
  code = FALSE,
  sp = NULL,
  nm = NULL,
  null.on.error = FALSE
)

get_fisheries(x, code = FALSE, fsh = NULL, nm = NULL, null.on.error = FALSE)

get_surveys(x, code = FALSE, sr = NULL, nm = NULL)
}
\arguments{
\item{conf}{The 'osmose.configuration' object.}

\item{par}{The name of the parameter, partial matching is allowed.}

\item{sp}{The number of the species (starting at 0).}

\item{fsh}{The number of the fishery (starting at 0).}

\item{invert}{Revert the selection (all but 'par')}

\item{as.is}{Boolean, TRUE if the results be returned as an osmose.configuration object, FALSE for a list or vector. 
The default is FALSE, but only when one exact match is found.}

\item{unlist}{Boolean, TRUE if you want the result to be returned as a vector instead of as a list.}

\item{linear}{Boolean, TRUE to transform 'log' and 'logit' parameters into linear scale.}

\item{type}{type of species to get the names for.}

\item{code}{Boolean, return the numerical code of the species or fishery?}
}
\value{
A list with all the matched parameters.
}
\description{
Get a parameter from a configuration file.
}
