% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose-class.R, R/osmose-main.R
\name{get_var.osmose}
\alias{get_var.osmose}
\alias{get_var.list}
\alias{get_var}
\title{Get variable from an \code{osmose}-like object.}
\usage{
\method{get_var}{osmose}(object, what, how = c("matrix", "list"), expected = FALSE, ...)

\method{get_var}{list}(object, what, how = c("matrix", "list"), expected = FALSE, ...)

get_var(object, what, how, ...)
}
\arguments{
\item{object}{Object of \code{osmose} class (see the \code{\link{read_osmose}} 
function).}

\item{what}{Name of variable to extract. See Details.}

\item{how}{How to return the object. Current options are "matrix" and "list".}

\item{expected}{A logical parameter. If \code{TRUE}, the average over the 
last dimensions will be performed (only if the output is an array).}

\item{...}{Additional arguments of the function.}
}
\value{
An array or a list containing the extracted data.
}
\description{
Function to get a variable from an object of \code{osmose} class.
This function uses the get_var method (see the \code{\link{get_var.osmose}}).
}
\details{
\code{what} can be any available variable contained on \code{object}
(e.g. biomass, abundance, yield, yieldN, etc).
}
