% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R, R/osmose-main.R
\name{getVar}
\alias{getVar}
\alias{get_var}
\title{Get variable from an \code{osmose}-like object.}
\usage{
getVar(object, what, how, ...)

get_var(object, what, how, ...)
}
\arguments{
\item{object}{Object of \code{osmose} class (see the \code{\link{read_osmose}} function).}

\item{what}{Variable to extract}

\item{how}{Output format}

\item{...}{Additional arguments of the function.}
}
\value{
An array or a list containing the extracted data.
}
\description{
Function to get a variable from an object of \code{osmose} class.
This function uses the get_var method (see the \code{\link{get_var.osmose}}).
}
