% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose-internal.R
\name{getOsmoseParameter}
\alias{getOsmoseParameter}
\title{Get a parameter from a name chain (error if not found).}
\usage{
getOsmoseParameter(x, ..., keep.att = FALSE)
}
\arguments{
\item{x}{Output of the \code{link{readOsmoseConfiguration}} function}

\item{...}{String arguments}

\item{keep.att}{Whether parameter attributes should be kept}
}
\description{
Get a parameter from a name chain (error if not found).
}
