% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose-auxiliar.R
\name{getSizeSpectrum}
\alias{getSizeSpectrum}
\title{Get size spectrum}
\usage{
getSizeSpectrum(file, sep = ",", ...)
}
\arguments{
\item{file}{File to read}

\item{sep}{File separator}

\item{...}{Additional arguments of the \code{read.csv} function}
}
\value{
A 3D array (time, length, species)
}
\description{
Get size spectrum
}
\examples{
{
dirin = system.file("extdata", package="osmose")
file = paste(dirin, "/outputs/SizeIndicators/gogosm_yieldDistribBySize_Simu0.csv", sep="")
size = getSizeSpectrum(file)
}
}
