% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose-class.R
\name{getVar.osmose}
\alias{getVar.osmose}
\title{GetVar method for osmose outputs objects}
\usage{
\method{getVar}{osmose}(object, what, how = c("matrix", "data.frame", "list"),
  expected = FALSE, ...)
}
\arguments{
\item{object}{Object of \code{osmose} class (see the \code{\link{read_osmose}} function).}

\item{what}{Name of variable to extract. It could be: "biomass","abundance",
"yield", "yieldN".}

\item{how}{How to return the object. Current options are "matrix" and "list".}

\item{expected}{A logical parameter. \code{TRUE} if the average over the 
last dimensions should be performed if the output is an array. By default is 
\code{expected = FALSE}.}

\item{...}{Additional arguments of the function.}
}
\value{
An array or a list containing the data.
}
\description{
Get a variable from an \code{osmose} object.
}
