% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose-class.R
\name{getVar}
\alias{getVar}
\title{Get variable}
\usage{
getVar(object, var, ...)
}
\arguments{
\item{object}{Object of \code{osmose} class (see the \code{\link{read_osmose}} function).}

\item{var}{Name of variable to extract. It could be: "biomass","abundance",
"yield", "yieldN".}

\item{...}{Additional arguments of the function.}
}
\value{
An array or a list containing the extracted data.
}
\description{
Function to get a variable from an object of \code{osmose} 
class. This function uses the getVar method (see the \code{\link{getVar.osmose}}).
}
