% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose_init.R
\name{initialize_osmose}
\alias{initialize_osmose}
\title{Create initialisation file for an OSMOSE configuration}
\usage{
initialize_osmose(
  input,
  file,
  type = "internannual",
  parameters = NULL,
  output = NULL,
  log = "osmose.log",
  version = NULL,
  osmose = NULL,
  java = "java",
  options = NULL,
  verbose = TRUE,
  clean = TRUE,
  force = FALSE,
  run = TRUE,
  append = FALSE,
  sp = NULL,
  ...
)
}
\arguments{
\item{input}{Filename of the main configuration file}

\item{file}{File to write the initialization configuration}

\item{type}{Initialisation type: currently 'interannual', 'climatological' and 'ncdf'}

\item{parameters}{Parameters to be passed to osmose (version 4 or higher).}

\item{output}{Output directory. If NULL, the value set in the configuration file is used.}

\item{log}{File to save OSMOSE execution messages.}

\item{version}{OSMOSE version. Integer (2, 3, etc.) or releases ('v3r2') are
accepted.}

\item{osmose}{Path to a OSMOSE .jar executable. By default (NULL), uses the stable
jar for the current version.}

\item{java}{Path to the java executable. The default assumes 'java' is
on the search path.}

\item{options}{Java options (e.g. -Xmx2048m to increase memory limit).}

\item{verbose}{Show messages? (output in the log file if FALSE).}

\item{clean}{TRUE if the output directory should be cleaned before running OSMOSE.}

\item{force}{TRUE if an outdated configuration can be run without being updated.}

\item{run}{Do we have to run the model to produce the outputs or are already there?}

\item{append}{To append or not the results to \code{file}}

\item{sp}{Species to initialise. NULL by default, meaning all species.}

\item{...}{Additional arguments, currently unused.}
}
\value{
The side effect is to create a configuration file with the initialization parameters.
}
\description{
Create initialisation file for an OSMOSE configuration
}
