\name{osmose-package}
\alias{osmose-package}
\alias{osmose}
\docType{package}
\title{
\packageTitle{osmose}
}
\description{
\packageDescription{osmose}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{osmose}
\packageIndices{osmose}
~~ An overview of how to use ~~
~~ the package, including the ~~
~~ most important functions ~~
}
\author{
\packageAuthor{osmose}

Maintainer: \packageMaintainer{osmose}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard ~~
~~ keywords, one per line, ~~
~~ from file KEYWORDS in the ~~
~~ R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
