% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/osmose-main.R
\docType{package}
\name{osmose-package}
\alias{osmose}
\alias{osmose-package}
\title{OSMOSE: Object-oriented Simulator of Marine biOdiverSity Exploitation}
\description{
\tabular{ll}{ Package: \tab osmose\cr Type: \tab Package\cr Version: \tab
0.1\cr Date: \tab 2016-02-01\cr License: \tab GPL-2\cr }
}
\examples{
osmose()
}
\author{
Yunne-Jai Shin

Philippe Verley

Morgane Travers

Laure Velez

Ricardo Oliveros-Ramos
Maintainer: Ricardo Oliveros-Ramos <ricardo.oliveros@gmail.com>
}
\references{
osmose: Modelling Marine Exploited Ecosystems
}
\keyword{ecosystems}
\keyword{marine}
\keyword{modelling}

