% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose-plot-util.R
\name{osmose.openfigure}
\alias{osmose.openfigure}
\title{Opens a figure file.}
\usage{
osmose.openfigure(figname, width = NULL, height = NULL)
}
\arguments{
\item{figname}{Name of the output file. The file sufix must
be .png or .pdf}

\item{width}{Figure width (default NULL)}

\item{height}{Figure height (default NULL)}
}
\description{
Opens a figure file.
}
