% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose-calibration.R
\name{osmose_calibration_outputs}
\alias{osmose_calibration_outputs}
\title{Process outputs for calibration from an OSMOSE object}
\usage{
osmose_calibration_outputs(output)
}
\arguments{
\item{output}{An object created by the \code{read_osmose} function.}
}
\value{
A list with the variables needed for a basic calibration of OSMOSE.
}
\description{
Process outputs for calibration from an OSMOSE object
}
\examples{
\dontrun{
output = read_osmose(".")
calib_output = osmose_calibration_outputs(output)
}
}
