% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose-calibration.R
\name{osmose_calibration_runmodel}
\alias{osmose_calibration_runmodel}
\title{Run a model as set up for a calibration trial}
\usage{
osmose_calibration_runmodel(
  input,
  osmose,
  name,
  version = "4.3.3",
  debug = FALSE
)
}
\arguments{
\item{input}{Filename of the main configuration file}

\item{osmose}{Path to a OSMOSE .jar executable. By default (NULL), uses the stable
jar for the current version.}

\item{name}{Name for the calibration trial.}

\item{version}{OSMOSE version. Integer (2, 3, etc.) or releases ('v3r2') are
accepted.}
}
\value{
A list with the simulated data used for calibration
}
\description{
Run a model as set up for a calibration trial
}
