% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose-calibration.R
\name{osmose_calibration_runmodel}
\alias{osmose_calibration_runmodel}
\title{Run a model as set up for a calibration trial}
\usage{
osmose_calibration_runmodel(
  input,
  name,
  version = "4.3.3",
  par = NULL,
  additional = NULL,
  debug = FALSE
)
}
\arguments{
\item{input}{Filename of the main configuration file}

\item{name}{Name for the calibration trial.}

\item{version}{OSMOSE version. Integer (2, 3, etc.) or releases ('v3r2') are
accepted.}

\item{par}{A list or osmose.configuration object containing OSMOSE parameters. Only calibrated parameters are accepted.}

\item{additional}{A list or osmose.configuration object containing OSMOSE parameters. Only NON calibrated parameters are accepted, particularly output parameters.}

\item{debug}{Do you want to debug the run_model function?}
}
\value{
A list with the simulated data used for calibration.
}
\description{
Run a model as set up for a calibration trial
}
