% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose-calibration.R
\name{osmose_calibration_setup}
\alias{osmose_calibration_setup}
\title{Set up a new calibration for an OSMOSE model}
\usage{
osmose_calibration_setup(
  input,
  osmose,
  name = NULL,
  data_path = NULL,
  type = "simple",
  bioen = FALSE,
  control = list(),
  version = "4.3.3",
  ...
)
}
\arguments{
\item{input}{Filename of the main configuration file}

\item{osmose}{Path to a OSMOSE .jar executable. By default (NULL), uses the stable
jar for the current version.}

\item{type}{Type or calibration. Currently only 'simple' is available.}

\item{bioen}{Are we calibrating a Bioen-OSMOSE model?}

\item{control}{A list with additional control arguments. Notably, control$dir to
choose the name of the calibration folder and control$run for the scratch folder 
(to 'run' OSMOSE during the calibration. Full paths are ignored, only basename is kept).}

\item{version}{OSMOSE version. Integer (2, 3, etc.) or releases ('v3r2') are
accepted.}

\item{...}{Additional arguments of current no use.}

\item{test}{Boolean, test the calibration? FALSE by default. You can run the same test later using
the function \code{\link{osmose_calibration_test}}.}
}
\value{
Side effects: Create all the files you need for your calibration. Returns, invisibily,
the path to the calibration directory.
}
\description{
Set up a new calibration for an OSMOSE model
}
