% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose_init-auxiliar.R
\name{osmose_grid}
\alias{osmose_grid}
\title{create OSMOSE ncdf grid (in development)}
\usage{
osmose_grid(conf, center = TRUE, ...)
}
\arguments{
\item{conf}{OSMOSE configuration}

\item{...}{}
}
\value{
An OSMOSE grid.
}
\description{
create OSMOSE ncdf grid (in development)
}
