% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose.config-class.R
\name{plot.osmose.config}
\alias{plot.osmose.config}
\title{Title}
\usage{
\method{plot}{osmose.config}(x, what = NULL, ...)
}
\arguments{
\item{x}{TODO}

\item{what}{TODO}

\item{...}{TODO}
}
\value{
TODO
}
\description{
Title
}
