% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose-plot.R
\name{plot.osmose.dietMatrix}
\alias{plot.osmose.dietMatrix}
\title{Plot the dietMatrix output}
\usage{
\method{plot}{osmose.dietMatrix}(
  x,
  time.mean = TRUE,
  species = NULL,
  thres = 1,
  add_text = TRUE,
  color = NULL,
  ...
)
}
\arguments{
\item{x}{osmose.dietMatrix output}

\item{time.mean}{True if time averaged should be drawn}

\item{species}{Species name (string or list). If NULL, al species are drawn.}

\item{thres}{Threshold (in percentage) below which preys are concatenated.}

\item{add_text}{True if values sjopu}

\item{color}{Histogram colors (if NULL, gray). Only used if time.mean=T}

\item{...}{Additional arguments to the barplot or line function}
}
\description{
Plot the dietMatrix output
}
