% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose-plot-inDevel.R
\name{plot.osmose.dietMatrixbyAge}
\alias{plot.osmose.dietMatrixbyAge}
\title{Plot diet matrix by size class}
\usage{
\method{plot}{osmose.dietMatrixbyAge}(x, species = NULL, ...)
}
\arguments{
\item{x}{Data list}

\item{species}{List of species name. If null, all species are plotted}

\item{...}{Extra arguments of the graphics barplot function}
}
\description{
Plot diet matrix by size class
}
