% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose-plot.R
\name{plot.osmose.mortalityRate}
\alias{plot.osmose.mortalityRate}
\title{Plot mortality rates}
\usage{
\method{plot}{osmose.mortalityRate}(x, time.mean = TRUE, species = NULL, norm = TRUE, ...)
}
\arguments{
\item{x}{Object of class osmose.mortalityRate}

\item{time.mean}{True if time-average plot should be ok}

\item{species}{List of species to plot. If NULL, all species are drawn.}

\item{norm}{True if percentage is returned instead of raw mort. rates}

\item{...}{Additional parameters to the barplot/lines functions}
}
\description{
Plot mortality rates
}
