% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose-plot-inDevel.R
\name{plot.osmose.predatorPressure}
\alias{plot.osmose.predatorPressure}
\title{Plots predator pressure.}
\usage{
\method{plot}{osmose.predatorPressure}(
  x,
  species = NULL,
  norm = TRUE,
  speciesNames = NULL,
  type = 1,
  thres = 0,
  parargs = list(),
  plotargs = list(),
  legargs = list(),
  axisargs = list(),
  draw_legend = TRUE,
  ...
)
}
\arguments{
\item{x}{Predator pressure.}

\item{species}{Species names}

\item{...}{Additional plot arguments}

\item{time.mean}{If TRUE, plots the time-average predator pressure 
sorted in decreasing order}
}
\description{
Plots predator pressure.
}
\details{
Thres > 0 does'nt work, since it returns negative values. To check what is going on.
}
