% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose-class.R
\name{plot.osmose}
\alias{plot.osmose}
\title{Plot Osmose outputs.}
\usage{
\method{plot}{osmose}(x, type = "biomass", ...)
}
\arguments{
\item{x}{Osmose outputs (see the \code{\link{osmose2R}} function)}

\item{type}{Data type to draw ("biomass", etc)}

\item{...}{Additional arguments of the plot function}
}
\value{
invisible()
}
\description{
Plot Osmose outputs.
}
