% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose-class.R
\name{print.osmose}
\alias{print.osmose}
\title{Print informations about Osmose outputs}
\usage{
\method{print}{osmose}(x)
}
\arguments{
\item{x}{Osmose outputs (see the \code{\link{osmose2R}} function)}
}
\description{
Print informations about Osmose outputs
}
