% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R, R/osmose-main.R
\name{osmose2R}
\alias{osmose2R}
\alias{read_osmose}
\title{Read OSMOSE outputs into an R object}
\usage{
osmose2R(path = NULL, version = "v3r2", species.names = NULL, ...)

read_osmose(
  path = NULL,
  input = NULL,
  version = "4.2.1",
  species.names = NULL,
  absolute = TRUE,
  ...
)
}
\arguments{
\item{path}{Path to the directory containing OSMOSE outputs.}

\item{version}{OSMOSE version used to run the model.}

\item{species.names}{Display names for species, overwrite the species names
provided to the OSMOSE model. Used for plots and summaries.}

\item{...}{Additional arguments}

\item{input}{Path to a main osmose configuration file.}

\item{absolute}{Whether the path is absolute (TRUE) or relative (FALSE). Only used if
input is not NULL}
}
\description{
This function create object of class \code{osmose} with the 
outputs from OSMOSE in the \code{path} folder.
}
\details{
A list of class \code{osmose} is created, individual elements can be
extracted using the function \code{get_var}.
}
\examples{
{
# Read outputs generated by demo (check osmose_demo function)
  outdir = system.file("extdata", "outputs/", package = "osmose")
  read_osmose(path = outdir)
}
}
\author{
Ricardo Oliveros-Ramos

Laure Velez
}
