% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose-class.R
\name{report.osmose}
\alias{report.osmose}
\title{Report method for osmose objects}
\usage{
\method{report}{osmose}(x, format = "latex", tangle = FALSE, output = NULL, open = TRUE)
}
\arguments{
\item{x}{Object of \code{osmose} class.}
}
\description{
Export a report of osmose objects.
}
