% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose-main.R
\name{report}
\alias{report}
\title{Report method}
\usage{
report(x, format, output, ...)
}
\arguments{
\item{format}{The format to export the report.}

\item{output}{Folder where the report will be saved.}

\item{...}{Extra arguments passed to \code{\link{report}} function.}

\item{object}{Object of class \code{osmose}.}
}
\value{
A report on specific format.
}
\description{
This function built a report for each class including on osmose package.
}
