% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose-main.R
\name{runOsmose}
\alias{runOsmose}
\title{Run an OSMOSE configuration}
\usage{
runOsmose(osmose = NULL, java = "java", input = "input/config.csv",
  output = "output/", options = NULL, log = "osmose.log",
  verbose = NULL, clean = TRUE, shell = "BASH")
}
\arguments{
\item{osmose}{Name of the Osmose java .jar file}

\item{java}{Name of the java executable}

\item{input}{Filename of the main configuration file}

\item{output}{Output directory}

\item{options}{Java options}

\item{log}{Log file}

\item{verbose}{NULL if run is not interactive (output in the log file)}

\item{clean}{TRUE if the output directory should be cleaned}

\item{shell}{Shell to use ("BASH" or "CSH")
a new file is created with the modified configuration.}
}
\description{
This function create a valid configuration by several input files
from user input parameters.
}
\details{
Basic configurations may not need the use of \code{buildConfiguration},
but it is required for configuration using interannual inputs or fishing selectivity.
}
\author{
Ricardo Oliveros-Ramos
}
