% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose-calibration-auxiliar.R
\name{set_par}
\alias{set_par}
\title{Set values for a parameter list}
\usage{
set_par(
  x,
  value = NULL,
  scale = 1,
  delta = 0,
  lower = NULL,
  upper = NULL,
  digits = 3,
  auto = FALSE,
  ...
)
}
\arguments{
\item{x}{The osmose.configuration list.}

\item{value}{The parameter value to replace.}

\item{scale}{All values in x are multiplied by this 'scale' units.}

\item{delta}{All values in x are shifted by this delta' units.}

\item{lower}{Minimum possible value after transformations (value and scale)}

\item{upper}{Maximum possible value after transformations (value and scale)}

\item{digits}{Maximum number of digits in the result}

\item{auto}{If auto=TRUE, scale is NOT applied to deviates and delta is ONLY applied to deviates.}

\item{...}{Additional arguments}
}
\value{
The configuration list with parameter values changed.
}
\description{
Set values for a parameter list
}
