% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose-calibration.R
\name{set_par}
\alias{set_par}
\title{Set values for a parameter list}
\usage{
set_par(x, value, ...)
}
\arguments{
\item{x}{The osmose.configuration list.}

\item{value}{The parameter value to replace.}

\item{...}{Additional arguments}
}
\value{
The configuration list with parameter values changed.
}
\description{
Set values for a parameter list
}
