% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose-calibration-auxiliar.R
\name{set_par}
\alias{set_par}
\title{Set values for a parameter list}
\usage{
set_par(
  x,
  value = NULL,
  scale = 1,
  lower = NULL,
  upper = NULL,
  delta = 0,
  digits = 3,
  ...
)
}
\arguments{
\item{x}{The osmose.configuration list.}

\item{value}{The parameter value to replace.}

\item{...}{Additional arguments}
}
\value{
The configuration list with parameter values changed.
}
\description{
Set values for a parameter list
}
