% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose.maps.R
\name{setOsmoseMaps}
\alias{setOsmoseMaps}
\title{Set maps of species distribution for OSMOSE}
\usage{
setOsmoseMaps(object, type, sp, lifespan, frequency = 24, normalize = TRUE,
  ...)
}
\arguments{
\item{object}{}

\item{type}{}

\item{sp}{}

\item{lifespan}{}

\item{frequency}{}

\item{normalize}{}

\item{interannual}{}

\item{start}{}

\item{end}{}

\item{ages}{}

\item{toPA}{}

\item{prob}{}

\item{criteria}{}

\item{lat}{}

\item{lon}{}
}
\description{
Set maps of species distribution for OSMOSE
}
