% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose-getVar.R
\name{summary.osmose.dietMatrix}
\alias{summary.osmose.dietMatrix}
\title{Title}
\usage{
\method{summary}{osmose.dietMatrix}(data, species = NULL, thres = 1,
  ...)
}
\arguments{
\item{data}{}

\item{species}{}

\item{thres}{}

\item{...}{}
}
\description{
Title
}
