% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose-getVar.R
\name{summary.osmose.mortalityRate}
\alias{summary.osmose.mortalityRate}
\title{Title}
\usage{
\method{summary}{osmose.mortalityRate}(data, species = NULL, ...)
}
\arguments{
\item{data}{}

\item{species}{}

\item{...}{}

\item{thres}{}
}
\description{
Title
}
