% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose-class.R
\name{summary.osmose}
\alias{summary.osmose}
\title{Summarizes informations about Osmose outputs}
\usage{
\method{summary}{osmose}(object)
}
\arguments{
\item{x}{Osmose outputs (see the \code{\link{osmose2R}} function)}
}
\description{
Summarizes informations about Osmose outputs
}
