% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R, R/osmose-auxiliar.R
\name{write.osmose}
\alias{write.osmose}
\alias{write_osmose}
\title{Write data in osmose format}
\usage{
write.osmose(x, file)

write_osmose(
  x,
  file,
  sep = ",",
  col.names = NA,
  quote = FALSE,
  row.names = TRUE,
  ...
)
}
\arguments{
\item{x}{Object to be written (table or data frame)}

\item{file}{Output file}

\item{sep}{The field separator string. Values within each row of x are 
separated by this string.}

\item{col.names}{either a logical value indicating whether the column names 
of x are to be written along with x, or a character vector of column names to 
be written. See the section on ‘CSV files’ for the meaning of 
\code{col.names = NA}.}

\item{quote}{A \code{logical} value (\code{TRUE} or \code{FALSE}) or a 
\code{numeric} vector.}

\item{row.names}{either a logical value indicating whether the row names of x 
are to be written along with x, or a character vector of row names to be 
written.}

\item{...}{Extra arguments passed to \code{write.table} funtion.}
}
\description{
Write an array or dataframe in the Osmose format.
The separator is ";", there are no quotes and a blank column is
added for the row names column.
}
