% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose-auxiliar.R
\name{write_osmose}
\alias{write_osmose}
\title{Write data in osmose format}
\usage{
write_osmose(x, file)
}
\arguments{
\item{x}{Object to be written (table or data frame)}

\item{file}{Output file}
}
\description{
Write an array or dataframe in the Osmose format.
The separator is ";", there are no quotes and a blank column is
added for the row names column.
}
